Solved a problem while closing TinyMCE in the item editor in FireFox
Pressing the close button in the back-end item editor in FireFox using TinyMCE resulted in a garbeld screen and a hang. Using other editors, i.e. JCE, CodeMirror on none, or using Chrome did yield this issue. Also, the problem was not available during 'save', 'save and close' and 'save and new'.

The problem was diagnosed to be related to unhiding TinyMCE before shutting the editor down. This code is triggered by the onBeforeSave event calling K2Editor->Save, but that code was not run for the close button.